% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{calcCumFrac}
\alias{calcCumFrac}
\title{A function to generate cumulative germination fractions from raw seed counts.}
\usage{
calcCumFrac(
  data,
  trt.id = "TrtID",
  cum.time = "CumTime",
  n.total = "nTotal",
  n.germinated = "nGerminated"
)
}
\arguments{
\item{data}{A dataset containing seed germination observations.}

\item{trt.id}{Column name containing ids which reflect the individual treatments.}

\item{cum.time}{Column name containing cumulative elapsted time.}

\item{n.total}{Total number of seeds for the given treatment.}

\item{n.germinated}{Column name containing total number of seeds germinated through the current time.}
}
\value{
a data frame with cumulative fractions generated from raw seed counts.
}
\description{
A function to generate cumulative germination fractions from raw seed counts.
}
\examples{
"foo"
}
\keyword{cumulative}
\keyword{fraction}
\keyword{germination}
\keyword{seed}
