% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTMsFunctions.R
\name{CalcHTModel}
\alias{CalcHTModel}
\title{A Function to calculate the Hydrotime model parameters.}
\usage{
CalcHTModel(Data, MaxCumFract)
}
\arguments{
\item{Data}{time course and cumulative dataset to be used in the hydrotime model. The original dataframe template should be used or column names should be modified similarly to the template. A column with time in hours CumTime) + a column with cumulative fractions (CumFract) and the experiment water potential (Germ.wp) are required. Filter the dataframe to only have treatments with water potential at the same temperature level.}

\item{MaxCumFract}{sets the ceiling cumulative fraction for the model when treatment at optimal condition displays a lower maximum cumulative fraction. Use it on your own discretion.}
}
\description{
This function calculates the hydrotime constant (HT), the median water potential base (Psib50) and the standard deviation (sigma).
}
\examples{
CalcHTModel(Data)
CalcHTModel(Data, MaxCumFract)
}
\keyword{hydrotime}
\keyword{model}
\keyword{parameters}
