% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priming_functions.R
\name{calcHTPModel}
\alias{calcHTPModel}
\title{A Function to calculate the Hydrothermal priming model parameters.}
\usage{
calcHTPModel(
  data,
  priming.wp = "PrimingWP",
  priming.temp = "PrimingTemp",
  priming.duration = "PrimingDuration",
  rate = "GR50",
  plot = TRUE
)
}
\arguments{
\item{data}{object with the calculated rates with treatments to be used in the Hydrothermal priming model. The output of the CalcSpeed function can be directly used here with the desired treatments. The fields with Treat.priming.wp, Treat.priming.temp and Treat.priming.duration need to be informed in the data file.}

\item{priming.wp}{Column containing the priming water potential treatments.}

\item{priming.temp}{Column containing the priming temperature treatments.}

\item{priming.duration}{Column containing the priming duration treatments.}

\item{rate}{Column containing the growth rate calculations from the calcSpeed function.}

\item{plot}{Should the model results be plotted?}
}
\description{
This function calculates the minimal water potential for priming effects (Psibmin50) and minimal temperature (Tmin).
}
\examples{
"foo"
}
\keyword{hydrothermal}
\keyword{model}
\keyword{parameters}
\keyword{priming}
