% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTMsFunctions.R
\name{CalcHTTModel}
\alias{CalcHTTModel}
\title{A Function to calculate the Hydrothermal time model parameters.}
\usage{
CalcHTTModel(Data, MaxCumFraction, Tb)
}
\arguments{
\item{Data}{time course and cumulative dataset to be used in the hydrothermal time model. The original dataframe template should be used or column names should be modified similarly to the template. A column with time (CumTime) + a column with cumulative fractions (CumFraction), the experimental water potential (Germ.wp) and the temperature (Germ.temp) are required.}

\item{MaxCumFraction}{sets the ceiling cumulative fraction for the model when treatment at optimal condition displays a lower maximum cumulative fraction. Use it on your own discretion.}

\item{Tb}{is the temperature base that can be provided and have a fixed value or left in blank to be calculated with other hydrothermal parameters.}
}
\description{
This function calculates the hydrohermal time constant (HT), the temperature base (tb), the median water potential base (Psib50) and the standard deviation (sigma).
}
\examples{
CalcHTTModel(mydata)
CalcHTTModel(mydata)
}
\keyword{hydrothermal}
\keyword{model}
\keyword{parameters}
\keyword{time}
