% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CalcSpeed}
\alias{CalcSpeed}
\title{A Speed function}
\usage{
CalcSpeed(
  Data,
  Fraction,
  T1ColName,
  T2ColName,
  T3ColName,
  T4ColName,
  T5ColName
)
}
\arguments{
\item{Data}{time course and cumulative dataset. A column with time in hours (Germ.time.hours) + a column with cumulative fractions (Germ.fraction) are required with at least one additional column for revelant treatment (e.g., germination temperature or water potential)}

\item{Fraction}{from 0 to 1 used to calculate the time required for that level to be obtained in the cumulative time course. Standard value is 0.5 (50%), to calculate the time to 50% germination (T50) and respective germination rate (GR50). Fraction level can be entered and be used for calculation and change column name.}

\item{T1ColName, T2ColName, T3ColName, T4ColName, T5ColName}{are the names of the treatment columns to separate the dataset. The time course cumulative curves will be grouped for each distinct treatment that should be informed here. These column names do not need to be informed in case the provided template file is used to organize the data.}
}
\description{
This function allows you to calculate the time to a desired germination fraction (Tx) and respective germination rate (GRx).
}
\examples{
CalcSpeed(MyData)
}
\keyword{GRx,}
\keyword{Tx,}
\keyword{germination}
\keyword{rate}
\keyword{speed,}
