% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTM-functions.R
\name{CalcT50nGR50}
\alias{CalcT50nGR50}
\title{A T50 and GR50 Function}
\usage{
CalcT50nGR50()
}
\description{
This function allows you to calculate the time to 50% germination (T50) and respective germination rate (GR50).
}
\examples{
CalcT50nGR50()
}
\keyword{GR50,}
\keyword{T50,}
\keyword{germination}
\keyword{rate}
\keyword{speed,}
