% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbtm_functions.R
\name{calcTTSubOModel}
\alias{calcTTSubOModel}
\title{A Function to calculate the ThermalTime model parameters. ------------------------}
\usage{
calcTTSubOModel(
  data,
  germ.temp = "GermTemp",
  cum.time = "CumTime",
  cum.frac = "CumFraction",
  max.cum.frac = 1,
  plot = TRUE
)
}
\arguments{
\item{data}{Time course and cumulative dataset to be used in the Thermaltime model. The original dataframe template should be used or column names should be modified similarly to the template. A column with time in hours (CumTime) + a column with cumulative fractions (CumFraction) and the experiment temperature (Germ.temp) are required. Filter the dataframe to only have treatments with temperature equal or under to  optimal temperature level.}

\item{germ.temp}{Column containing germination temperature values.}

\item{cum.time}{Column containing cumulative elapsed time.}

\item{cum.frac}{Column containing cumulative fraction germinated.}

\item{max.cum.frac}{Sets the ceiling cumulative fraction for the model when treatment at optimal condition displays a lower maximum cumulative fraction. Use it on your own discretion.}

\item{plot}{Should the model results be plotted?}
}
\description{
This function calculates the temperature base (Tb), the thermal time to 50\% of the population (ThetaT50) and the standard deviation (sigma).
}
\keyword{Thermal}
\keyword{model}
\keyword{parameters}
\keyword{time}
