% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{cleanData}
\alias{cleanData}
\title{A Function to clean cumulative curves on dataset.}
\usage{
cleanData(data, treatments = c("TrtID"), cum.frac = "CumFraction")
}
\arguments{
\item{data}{A data frame object with the raw cumulative data that needs to be cleaned.}

\item{treatments}{are the names of the treatment columns to separate the dataset. The time course cumulative curves will be grouped for each distinct treatment that should be informed here.}

\item{cum.frac}{Column name for cumulative fraction germinated.}
}
\value{
A data frame with duplicate observations removed
}
\description{
This function removes repetitive cumulative fractions/percentages, keeping only the initial presence of the value
}
\examples{
"foo"
}
\keyword{clean}
\keyword{cumulative}
\keyword{fraction}
\keyword{percentage}
\keyword{repetitive}
