% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTM-functions.R
\name{DefineMaxGerm}
\alias{DefineMaxGerm}
\title{A Function to get maximum germination percentage on dataset.}
\usage{
DefineMaxGerm(SeedLot, GermWP, GermTEMP)
}
\description{
This function normalizes all data to the percentage informed here. This is useful when a known fraction of the population is not viable.
}
\examples{
DefineMaxGerm()
DefineMaxGerm()
}
\keyword{Maximum}
\keyword{germination}
