% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTM-functions.R
\name{DefineModel}
\alias{DefineModel}
\title{A Function to define the population-based threshold model(PBTM) you want to work with.}
\usage{
DefineModel(ModelCode)
}
\arguments{
\item{DefineModel(1)}{selects the Hydropriming Model}

\item{DefineModel(1.5)}{selects the Hydrothernal priming Model}

\item{DefineModel(2)}{selects the Suboptimal Hydrotime Model}
}
\value{
Sets up the package and print the selected model.
}
\description{
This function allows you to choose the PBT model to be calculated. Choose the model that you want to work on (1)Hydropriming model; (2)Suboptimal Hydrotime; (2.5)Supra-optimal Hydrotime; (3)Thermaltime; (4)Suboptimal Hydrothermal Time; (5)Supra-optimal Hydrothermal Time.
}
\examples{
DefineModel(1.5)
}
\keyword{PBTM}
