% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTM-functions.R
\name{FixedTo}
\alias{FixedTo}
\title{A Function to manually define the optimal temperature (To) for model calculation.}
\usage{
FixedTo(To1)
}
\description{
This function defines the To value and avoid it to be calculated together with other values.
}
\examples{
FixedTo(22.5)
FixedTo()
}
\keyword{Known}
\keyword{To}
\keyword{fixed}
\keyword{otpimal}
\keyword{temperature}
