% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{mergeTrts}
\alias{mergeTrts}
\title{A function to regenerate cumulative fractions by merging treatments.}
\usage{
mergeTrts(
  data,
  keep.trts,
  trt.id = "TrtID",
  cum.time = "CumTime",
  cum.frac = "CumFraction"
)
}
\arguments{
\item{data}{A data frame object with the raw cumulative data that needs to be cleaned.}

\item{keep.trts}{One or more treatments to keep from the original set of treatments.}

\item{trt.id}{The column specifying the original set of treatments used to generate the cumulative fractions.}

\item{cum.time}{Column name for the cumulative time data.}

\item{cum.frac}{Column name for the cumulative fraction germinated.}
}
\value{
A data frame with new cumulative fraction values.
}
\description{
A function to regenerate cumulative fractions by merging treatments.
}
\examples{
"foo"
}
\keyword{cumulative}
\keyword{fraction}
\keyword{merge}
\keyword{treatments}
