% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myGermData.R
\docType{data}
\name{myGermData}
\alias{myGermData}
\title{Sample seed germination data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 398 rows and 7 columns.
}
\usage{
myGermData
}
\description{
Sample seed germination data
}
\keyword{datasets}
