% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPrimingData.R
\docType{data}
\name{myPrimingData}
\alias{myPrimingData}
\title{Sample seed priming data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 259 rows and 10 columns.
}
\usage{
myPrimingData
}
\description{
Sample seed priming data
}
\keyword{datasets}
