% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priming_plots.R
\name{plotHPModel}
\alias{plotHPModel}
\title{A Function to plot the both priming models.}
\usage{
plotHPModel(
  data,
  model,
  priming.wp = "PrimingWP",
  priming.duration = "PrimingDuration"
)
}
\arguments{
\item{data}{object with the calculated rates with treatments to be used in the Hydrothermal priming model. The output of the CalcSpeed function can be directly used here with the desired treatments. The fields with Treat.priming.wp, Treat.priming.temp and Treat.priming.duration need to be informed in the data file.}

\item{model}{is data object resulting from the CalcHPModel() or CalcHTPModel() functions containing the model information and parameter results.}

\item{priming.wp}{Column name for the priming water potential.}

\item{priming.duration}{Column name for the duration of priming.}
}
\description{
This function plots the priming models and calculated parameters.
}
\examples{
"foo"
}
\keyword{hydropriming}
\keyword{hydrothermal}
\keyword{model}
\keyword{plot}
\keyword{priming}
