% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priming_plots.R
\name{plotHTPModel}
\alias{plotHTPModel}
\title{A Function to plot the both priming models.}
\usage{
plotHTPModel(
  data,
  model,
  priming.wp = "PrimingWP",
  priming.temp = "PrimingTemp",
  priming.duration = "PrimingDuration"
)
}
\arguments{
\item{data}{object with the calculated rates with treatments to be used in the Hydrothermal priming model. The output of the CalcSpeed function can be directly used here with the desired treatments.}

\item{model}{is data object resulting from the calcHPModel or calcHTPModel functions containing the model information and parameter results.}

\item{priming.wp}{Column name for the priming water potential.}

\item{priming.temp}{Column name for the priming temperature.}

\item{priming.duration}{Column name for the duration of priming.}
}
\description{
This function plots the priming models and calculated parameters.
}
\keyword{hydropriming}
\keyword{hydrothermal}
\keyword{model}
\keyword{plot}
\keyword{priming}
