% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PlotModl}
\alias{PlotModl}
\title{A Function to plot the selected calculated model and parameters and predicitions.}
\usage{
PlotModl()
}
\description{
This function plots the selected model and calculated parameters.
}
\examples{
PlotModl()
PlotModl()
}
\keyword{model}
\keyword{plot}
\keyword{population-based}
\keyword{threshold}
