% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTMsPlots.R
\name{PlotPBTMModel}
\alias{PlotPBTMModel}
\title{A Function to plot the selected calculated model and parameters and predicitions.}
\usage{
PlotPBTMModel(Data, ModelResults)
}
\arguments{
\item{Data}{time course and cumulative dataset to be used in the Thermaltime model. The original dataframe template should be used or column names should be modified similarly to the template. A column with time in hours (CumTime) + a column with cumulative fractions (CumFraction) and the experiment temperature (Germ.temp) are required. Filter the dataframe to only have treatments with temperature equal or under to  optimal temperature level.}

\item{ModelResults}{is data object resulting from the Calc PBTM functions containing the model information and parameter results.}
}
\description{
This function plots the selected model and calculated parameters.
}
\examples{
PlotPBTMModel(mydata, myModelResults)
PlotPBTMModel(mydata, myModelResults)
}
\keyword{model}
\keyword{plot}
\keyword{population-based}
\keyword{threshold}
