% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrimingModelsPlots.R
\name{PlotPrimingModel}
\alias{PlotPrimingModel}
\title{A Function to plot the both priming models.}
\usage{
PlotPrimingModel(Data, ModelResults, GR)
}
\arguments{
\item{ModelResults}{is data object resulting from the CalcHPModel() or CalcHTPModel() functions containing the model information and parameter results.}

\item{GR}{is the column name for the rate to be used in the model and needs to be informed in case the data object contains a GR different than GR50.}

\item{myData}{object with the calculated rates with treatments to be used in the Hydrothermal priming model. The output of the CalcSpeed function can be directly used here with the desired treatments. The fields with Treat.priming.wp, Treat.priming.temp and Treat.priming.duration need to be informed in the data file.}
}
\description{
This function plots the priming models and calculated parameters.
}
\examples{
PlotPrimingModel(myData, HPModelResults)
PlotPrimingModel(myData, HPModelResults)
}
\keyword{hydropriming}
\keyword{hydrothermal}
\keyword{model}
\keyword{plot}
\keyword{priming}
