% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PlotRateVsTreat}
\alias{PlotRateVsTreat}
\title{A Function to plot rate vs treatment temperature.}
\usage{
PlotRateVsTreat(Data, x, y)
}
\description{
This function plots rates against the desired treatment.
}
\examples{
PlotRateVsTreat()
PlotRateVsTreat()
}
\keyword{GR50}
\keyword{Temperature}
\keyword{plot}
