% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_plots.R
\name{plotRateVsTrt}
\alias{plotRateVsTrt}
\title{A Function to plot rate vs desired treatment.}
\usage{
plotRateVsTrt(data, x = "TrtID", y = "GR50")
}
\arguments{
\item{data}{A data frame generated by the calsSpeed function. It should have the summarized treatments and respective GR values.}

\item{x}{should indicate the treatment column name for the x axis (e.g., "GermTemp", "GermWP" or others).}

\item{y}{should indicate the rate column name for the y axis if different than GR50 (e.g., "GR10", "GR90", etc).}
}
\description{
This function plots rates against the desired treatment.
}
\examples{
"foo"
}
\keyword{Temperature}
\keyword{plot}
\keyword{rates}
