% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTM-functions.R
\name{PlotRawData}
\alias{PlotRawData}
\title{A Function to plot the raw data selected.}
\usage{
PlotRawData()
}
\arguments{
\item{PlotRawData()}{will plot raw cumulative curves over time.}
}
\description{
This function plots the raw data previously selected.
}
\examples{
PlotRawData()
PlotRawData()
}
\keyword{data}
\keyword{plot}
\keyword{raw}
