% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbtm_plots.R
\name{plotTTSubOModel}
\alias{plotTTSubOModel}
\title{A Function to plot the selected calculated model and parameters and predicitions.}
\usage{
plotTTSubOModel(
  data,
  model,
  germ.temp = "GermTemp",
  cum.time = "CumTime",
  cum.frac = "CumFraction"
)
}
\arguments{
\item{data}{A data frame containing time course and cumulative germination fractions to be used in the ThermalTime model. A column with time in hours, a column with cumulative fractions, and the experiment temperature are required.}

\item{model}{is the results list returned from from running any of the PBT models.}

\item{germ.temp}{Name of the column for the experimental temperature.}

\item{cum.time}{Name of the column for cumulative time.}

\item{cum.frac}{Name of the the column for cumulative fraction germinated.}
}
\description{
This function plots the selected model and calculated parameters.
}
\keyword{hydrotime}
\keyword{model}
\keyword{plot}
\keyword{population-based}
\keyword{threshold}
