% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_priming_data.R
\docType{data}
\name{sample_priming_data}
\alias{sample_priming_data}
\title{Sample seed priming data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 259 rows and 10 columns.
}
\usage{
sample_priming_data
}
\description{
Sample seed priming data
}
\keyword{datasets}
