% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priming_functions.R
\name{calcHPModel}
\alias{calcHPModel}
\title{A Function to calculate the Hydropriming model parameters.}
\usage{
calcHPModel(
  data,
  priming.wp = "PrimingWP",
  priming.duration = "PrimingDuration",
  rate = "GR50",
  plot = TRUE
)
}
\arguments{
\item{data}{Data rame with the calculated rates with treatments to be used in the Hydropriming model. The output of the calcSpeed function can be directly used here with the desired treatments. Priming water potential and priming duration treatments are required.}

\item{priming.wp}{Column containing the priming water potential treatments.}

\item{priming.duration}{Column containing the priming duration treatments.}

\item{rate}{Column containing the growth rate data from the calcSpeed function.}

\item{plot}{Should the model results be plotted?}
}
\description{
This function calculates the minimal water potential for priming effects (Psib50min).
}
\examples{
"foo"
}
\keyword{hydropriming}
\keyword{model}
\keyword{parameters}
