% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbtm_functions.R
\name{calcHTTModel}
\alias{calcHTTModel}
\title{A Function to calculate the Hydrothermal time model parameters.}
\usage{
calcHTTModel(
  data,
  germ.wp = "GermWP",
  germ.temp = "GermTemp",
  cum.time = "CumTime",
  cum.frac = "CumFraction",
  max.cum.frac = 1,
  base.temp = NULL,
  plot = TRUE
)
}
\arguments{
\item{data}{time course and cumulative dataset to be used in the hydrothermal time model. The original dataframe template should be used or column names should be modified similarly to the template. A column with time (CumTime) + a column with cumulative fractions (CumFraction), the experimental water potential (Germ.wp) and the temperature (Germ.temp) are required.}

\item{germ.wp}{Column containing germination water potential.}

\item{germ.temp}{Column containing germination temperature.}

\item{cum.time}{Column containing cumulative elapsed time.}

\item{cum.frac}{Column containin cumulative fraction germinated.}

\item{max.cum.frac}{Sets the ceiling cumulative fraction for the model when treatment at optimal condition displays a lower maximum cumulative fraction. Use it on your own discretion.}

\item{base.temp}{is the temperature base that can be provided and have a fixed value or left in blank to be calculated with other hydrothermal parameters.}

\item{plot}{Should the model results be plotted?}
}
\description{
This function calculates the hydrohermal time constant (HT), the temperature base (tb), the median water potential base (Psib50) and the standard deviation (sigma).
}
\keyword{hydrothermal}
\keyword{model}
\keyword{parameters}
\keyword{time}
