% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CalcT50nGR50}
\alias{CalcT50nGR50}
\title{A T50 and GR50 Function}
\usage{
CalcT50nGR50(
  Data,
  T1ColName,
  T2ColName,
  T3ColName,
  T4ColName,
  T5ColName,
  TimeColName,
  CumFractColName
)
}
\arguments{
\item{Data}{time course and cumulative dataset. A column with time in hours (Germ.time.hours) + a column with cumulative fractions (Germ.fraction) are required with at least one additional column for revelant treatment (e.g., germination temperature or water potential)}

\item{T1ColName, T2ColName, T3ColName, T4ColName, T5ColName}{are the names of the treatment columns to separate the dataset. The time course cumulative curves will be grouped for each distinct treatment that should be informed here. These column names do not need to be informed in case the provided template file is used to organize the data.}

\item{TimeColName, CumFractColName}{are the names of the data columns that will be used to calculate values. In the case the column names for time (Germ.time.hours) and cumulative fraction (Germ.fraction) used are equal to the template provided, these column names to not need to be provided.}
}
\description{
This function allows you to calculate the time to 50% germination (T50) and respective germination rate (GR50).
}
\examples{
CalcT50nGR50(MyData)
}
\keyword{GR50,}
\keyword{T50,}
\keyword{germination}
\keyword{rate}
\keyword{speed,}
