% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{CleanData}
\alias{CleanData}
\title{A Function to clean cumulative curves on dataset.}
\usage{
CleanData(Data, Treat1, Treat2, Treat3, Treat4, Treat5)
}
\arguments{
\item{Data}{object with the raw cumulative data that needs to be removed.}

\item{Treat1, Treat2, Treat3, Treat4, Treat5}{are the names of the treatment columns to separate the dataset. The time course cumulative curves will be grouped for each distinct treatment that should be informed here.}
}
\description{
This function removes repetitive cumulative fractions/percentages, keeping only the initial presence of the value
}
\examples{
CleanData(mydata,"Treat.desc")
CleanData(mydata,"Treat.desc")
}
\keyword{Clean}
\keyword{cumulative}
\keyword{fraction}
\keyword{percentage}
\keyword{repetitive}
