% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTM-functions.R
\name{DefineHTo}
\alias{DefineHTo}
\title{A Function to manually define the Hydrotime constant at optimal temperature (HTo) for model calculation.}
\usage{
DefineHTo(HTo1)
}
\arguments{
\item{DefineHTo(50)}{will define hydrotime constant to 50 for further supra-optimal hydrotime model calculation.}
}
\description{
This function defines the HTo value and avoid it to be calculated together with other values.
}
\examples{
DefineHTo(50)
DefineHTo()
}
\keyword{HTo}
\keyword{Hydrotime}
\keyword{Known}
\keyword{constant}
\keyword{fixed}
\keyword{hydrotime}
\keyword{supra-optimal}
\keyword{temperature}
