% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PlotRateVsTreat}
\alias{PlotRateVsTreat}
\title{A Function to plot rate vs treatment temperature.}
\usage{
PlotRateVsTreat(Data, x, y)
}
\arguments{
\item{Data}{should inform the table with time course data to be plotted}

\item{x}{should indicate the treatment column name for the x axis.}

\item{y}{should indicate the rate column name for the y axis if different than GR50.}
}
\description{
This function plots rates against the desired treatment.
}
\examples{
PlotRateVsTreat()
PlotRateVsTreat()
}
\keyword{GR50}
\keyword{Temperature}
\keyword{plot}
