% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{PlotRateVsTreat}
\alias{PlotRateVsTreat}
\title{A Function to plot rate vs desired treatment.}
\usage{
PlotRateVsTreat(Data, x, y)
}
\arguments{
\item{Data}{should inform the table that resulted from the CalcSpeed function. It should have the summarized treatments and respective GR values.}

\item{x}{should indicate the treatment column name for the x axis (e.g., "Germ.temp", "Germ.wp" or others).}

\item{y}{should indicate the rate column name for the y axis if different than GR50 (e.g., "GR90", "GR10", etc).}
}
\description{
This function plots rates against the desired treatment.
}
\examples{
PlotRateVsTreat(MyCalcSpeedData,"Germ.temp")
PlotRateVsTreat(MyCalcSpeedData,"Germ.temp")
}
\keyword{Temperature}
\keyword{plot}
\keyword{rates}
