% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_plots.R
\name{plotRawData}
\alias{plotRawData}
\title{A Function to plot the cumulative raw data selected.}
\usage{
plotRawData(
  data,
  color = NULL,
  shape = NULL,
  line = TRUE,
  cum.time = "CumTime",
  cum.frac = "CumFraction"
)
}
\arguments{
\item{data}{A data frame with time course data for one or two treatments.}

\item{color}{The treatment to be mapped to color on the plot.}

\item{shape}{The treatment to be mapped to shape on the plot.}

\item{line}{Should the line overlay be added?}

\item{cum.time}{Column name for cumulative elapsed time.}

\item{cum.frac}{Column name for cumulative fraction germinated.}
}
\description{
This function plots the raw data previously selected. It will plot raw cumulative curves over time.
}
\keyword{data}
\keyword{plot}
\keyword{raw}
