% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PlotRawDt}
\alias{PlotRawDt}
\title{A Function to plot the cumulative raw data selected.}
\usage{
PlotRawDt(Data, Treat1, Treat2, Treat3)
}
\arguments{
\item{Data}{with time course data for one or more treatments.}

\item{Treat1, Treat2, Treat3}{are the desired treatment factors to be informed as column names. The first informed treatment will be separated as color, the second as shape and third as alpha (transparency).}
}
\description{
This function plots the raw data previously selected. It will plot raw cumulative curves over time.
}
\examples{
PlotRawDt(MyData)
PlotRawDt(MyRawData)
}
\keyword{data}
\keyword{plot}
\keyword{raw}
